dev variables established, further reorganization
