+++ capture/devconfig.py
@@ -1,0 +1,16 @@
+# devconfig.py
+
+#user target file
+rxnvarfile = "WF1_template.xlsx"
+
+#version control
+RoboVersion = 2.1
+
+# Hard coded limits
+max_robot_reagents = 7
+
+
+
+
+
+
