+++ capture/inspect/plotter.py
