+++ capture/inspect/plotter_broke.py
